SymbolicFuzzer: Solve import errors
In this project, we need to use multiple packages from fuzzingbook.
However, since we might need to make some changes to SymbolicFuzzer package, we need to make
some changes to the import statements.

PS: We may delete the local file SymbolicFuzzer.py if the final design does not need it.