Support of unsaturated cores
In order to get unsaturated cores printed, each constraint must be annotated with a z3.Bool variables and a name.
So we had to create those booleans dynamically according to the number of constraints in each statement, then
we name them the same as the content of the constraints. Hence, unsat_core() will print the source of unsat core.